SD - Move SDCard class into own header file
